Docker compose: Flytter override urler ut til common.env (#852)
* Docker compose: Flytter override urler ut til common.env

* Fjerner endring av setup

* Rydder litt i setup-lokal-utvikling.sh

* Rydder litt mer i setup-lokal-utvikling.sh

+++ lokal-utvikling/setup-lokal-utvikling.sh
@@ -70,20 +70,29 @@ settPorterSomSkalErstattes () {
   esac
 }
 
+host_adresse="host.docker.internal"
 while [ -n "$1" ]; do # while loop starts
-
     case "$1" in
       -m|--mock)
         mock_applikasjon+=("$2")
+        shift
+        ;;
+      -s|--system)
+        if [[ "$2" =~ "colima"  ]] ; then
+            host_adresse="192.168.5.2"
+        fi
         shift
         ;;
       --help)
         echo "usage: ./setup-lokal-utvikling.sh [options] [APPLIKASJON_UTENFOR_DOCKER_COMPOSE ...]"
         echo ""
         echo "Options:"
-        echo "-m,--mock <applikasjon>     Her kan du velge å mocke ut spesifikke applikasjoner istedenfor å bruke den"
-        echo "                            faktiske applikasjonen. En mock av applikasjonen i VTP blir dermed brukt."
-        echo "                            Applikasjoner som kan mockes ut er fptilbake, fpoppdrag og fpformidling."
+        echo "-s, --system <docker_desktop|colima>      Velg enten docker_desktop eller colima. Brukes bare av colima brukere"
+        echo "                                          default: docker_desktop"
+        echo "-m, --mock <applikasjon>                  Her kan du velge å mocke ut spesifikke applikasjoner istedenfor å bruke den"
+        echo "                                          faktiske applikasjonen. En mock av applikasjonen i VTP blir dermed brukt."
+        echo "                                          Applikasjoner som kan mockes ut er fptilbake, fpoppdrag og fpformidling."
+        echo "                                          default: ingen applikasjoner mockes ut"
         exit 0
         ;;
       *)
@@ -91,7 +100,6 @@ while [ -n "$1" ]; do # while loop starts
         ;;
     esac
     shift
-
 done
 
 
@@ -103,6 +111,7 @@ if [ $# -eq 0 ]; then
   echo "Applikasjonnavn er ikke spesifisert. Lager dermed oppsett for å kjøre den minste verdikjeden i Docker Compose."
 else
   echo "Applikasjonen(e) som må startes opp i IDE er: ${applikasjoner[*]}"
+  echo "Containernavn for ${applikasjoner[*]} erstattes med ${host_adresse}"
 fi
 
 if [ -d "$folder" ]; then
@@ -111,47 +120,59 @@ fi
 mkdir $folder
 
 
+# Copy all docker-compose file to separate docker-compose-lokal folder for local instance
 cd ../resources/pipeline
 settPorterSomSkalErstattes ${applikasjoner[0]}
 for f in {.*,*}; do
     if [[ -f "$f" ]]; then
-      cp "$f" "${relativ_path}/$f"
+        cp "$f" "${relativ_path}/$f"
     elif [[ $f == oracle-init ]] || [[ $f == postgres-init ]] || [[ $f == tokenx ]]; then
-      cp -r "$f" "${relativ_path}"
+        cp -r "$f" "${relativ_path}"
     fi
 done
 
-
+# Redirect alle kall for angitte applikasjoner lokalt på maskinen din. F.eks. ved valg av fpabonnent erstattes
+# http://fpabonnent:8080/fpabonnent med http://host.docker.internal:8065/fpabonnent
 cd $relativ_path
 for applikasjon in "${applikasjoner[@]}"; do
   settPorterSomSkalErstattes $applikasjon
   for f in {.*,*}; do
-    if [ -f "$f" ] && [[ $f != .env ]]  && [[ $f != *.sh ]]; then
+    if [ -f "$f" ] && [[ $f != .env ]]  && [[ $f != *.sh ]] && [[ $f != *.sql ]]; then # Only files && not .env %% not script files && not sql files
       for ((i=0;i<${#replace_port_array[@]};++i)); do
-        sed -i.bak "s/${applikasjon}:${replace_port_array[i]}/host.docker.internal:${with_port_array[i]}/g" "$f"
+        sed -i.bak "s/${applikasjon}:${replace_port_array[i]}/${host_adresse}:${with_port_array[i]}/g" "$f"
         rm $f.bak
       done
     fi
   done
 done
 
-applikasjoner_som_kan_mockes_ut=(fpoppdrag fptilbake fpformidling fprisk)
-echo "Mocker ut følgende applikasjoner: ${mock_applikasjon[@]}"
-for applikasjon in "${mock_applikasjon[@]}"; do
-  if [[ "${applikasjoner_som_kan_mockes_ut[@]}" =~ "${applikasjon}" ]]; then
-    if [[ "$*" == *vtp* ]]; then
-      sed -i.bak "s*${applikasjon}:8080*host.docker.internal:8060/rest/dummy*g" "docker-compose.yml"
-    else
-      sed -i.bak "s*${applikasjon}:8080*vtp:8060/rest/dummy*g" "docker-compose.yml"
-    fi
-    if [[ "${applikasjon}" =~ "fpoppdrag"  ]] && [[ "$*" == *fpfrontend* ]] ; then
-      sed -i.bak "s*localhost:9000/fpoppdrag/api*vtp:8060/rest/dummy/fpoppdrag/api*g" "docker-compose.yml"
-    fi
-    rm docker-compose.yml.bak
-  else
-    echo "Mock av ${applikasjon} finnes ikke – beholder konfigurasjon som den er."
-  fi
-done
+# Ved mocking av våre interne applikasjoner bruker vi en stub i vtp. Bare støtte for fpoppdrag, fptilbake, fpformidling eller fprisk
+if [ ${#mock_applikasjon[@]} ]; then
+    echo "Mocker ut følgende applikasjoner: ${mock_applikasjon[@]}"
+    applikasjoner_som_kan_mockes_ut=(fpoppdrag fptilbake fpformidling fprisk)
+    for applikasjon in "${mock_applikasjon[@]}"; do
+        if [[ "${applikasjoner_som_kan_mockes_ut[@]}" =~ "${applikasjon}" ]]; then
+            if [[ "$*" == *vtp* ]]; then
+                sed -i.bak "s*${applikasjon}:8080*${host_adresse}:8060/rest/dummy*g" "docker-compose.yml"
+                sed -i.bak "s*${applikasjon}:8080*${host_adresse}:8060/rest/dummy*g" "common.env"
+            else
+                sed -i.bak "s*${applikasjon}:8080*vtp:8060/rest/dummy*g" "docker-compose.yml"
+                sed -i.bak "s*${applikasjon}:8080*vtp:8060/rest/dummy*g" "common.env"
+            fi
+
+            if [[ "${applikasjon}" =~ "fpoppdrag"  ]] && [[ "$*" == *fpfrontend* ]] ; then
+                sed -i.bak "s*localhost:9000/fpoppdrag/api*vtp:8060/rest/dummy/fpoppdrag/api*g" "docker-compose.yml"
+                sed -i.bak "s*localhost:9000/fpoppdrag/api*vtp:8060/rest/dummy/fpoppdrag/api*g" "common.env"
+            fi
+
+            rm docker-compose.yml.bak
+            rm common.env.bak
+        else
+            echo "Mock av ${applikasjon} finnes ikke – beholder konfigurasjon som den er"
+        fi
+    done
+fi
+
 
 cd ../../resources/pipeline
 cp "update-versions.sh" "${relativ_path}"
